move NotificationError to error.rs
